Rename to rtsp_to_webrtc to follow standards
