+++ mypy.ini
@@ -1298,7 +1298,7 @@ no_implicit_optional = true
 warn_return_any = true
 warn_unreachable = true
 
-[mypy-homeassistant.components.rtsptowebrtc.*]
+[mypy-homeassistant.components.rtsp_to_webrtc.*]
 check_untyped_defs = true
 disallow_incomplete_defs = true
 disallow_subclassing_any = true
