+++ requirements_all.txt
@@ -2110,7 +2110,7 @@ rpi-bad-power==0.1.0
 # homeassistant.components.rpi_rf
 # rpi-rf==0.9.7
 
-# homeassistant.components.rtsptowebrtc
+# homeassistant.components.rtsp_to_webrtc
 rtsp-to-webrtc==0.2.7
 
 # homeassistant.components.russound_rnet
