+++ requirements_test_all.txt
@@ -1275,7 +1275,7 @@ roonapi==0.0.38
 # homeassistant.components.rpi_power
 rpi-bad-power==0.1.0
 
-# homeassistant.components.rtsptowebrtc
+# homeassistant.components.rtsp_to_webrtc
 rtsp-to-webrtc==0.2.7
 
 # homeassistant.components.yamaha
